moves handle_metadata() into TasksStorage
- moves the handle_metadata() function into the implementation block of
TasksStorage.
- changes ensure_valid_vote() to use the new method.
- Moves the Verifier helper errors into APIs so they can be reused
throughout the project.
