+++ packages/apis/src/interfaces/tasks.rs
@@ -2,11 +2,19 @@
 /// by other contracts. This specifies the required APIs for on-chain interactions.
 /// This must be a subset of any of the implementation.
 use cosmwasm_schema::{cw_serde, QueryResponses};
-use cosmwasm_std::{Addr, StdResult, Storage, Timestamp};
+use cosmwasm_std::{Addr, Decimal, DepsMut, Env, StdResult, Storage, Timestamp};
 use cw_orch::{ExecuteFns, QueryFns};
 use cw_storage_plus::Map;
 
-use crate::{id::TaskId, tasks::Status, verifier_simple::TaskMetadata};
+use crate::{
+    id::TaskId,
+    tasks::Status,
+    verifier_simple::{helper::VerifierError, TaskMetadata},
+};
+
+use crate::interfaces::voting::QueryMsg as OperatorQueryMsg;
+
+use super::voting::TotalPowerResponse;
 
 // FIXME: hot to make these generic
 pub type ResponseType = serde_json::Value;
@@ -37,6 +45,68 @@ impl<'a> TasksStorage<'a> {
     ) -> StdResult<()> {
         self.0.save(store, key, &metadata)?;
         Ok(())
+    }
+
+    pub fn handle_metadata(
+        deps: DepsMut,
+        env: &Env,
+        operator_addr: &Addr,
+        task_queue: &Addr,
+        task_id: TaskId,
+        fraction_percent: u32,
+    ) -> Result<TaskMetadata, VerifierError> {
+        let tasks_storage = TasksStorage::new("tasks");
+        let metadata = tasks_storage.get_tasks(deps.storage, (task_queue, task_id))?;
+
+        match metadata {
+            Some(meta) => {
+                // Ensure this is not yet expired (or completed)
+                match meta.status {
+                    TaskStatus::Completed => Err(VerifierError::TaskAlreadyCompleted),
+                    TaskStatus::Expired => Err(VerifierError::TaskExpired),
+                    TaskStatus::Open if meta.is_expired(env) => Err(VerifierError::TaskExpired),
+                    _ => Ok(meta.clone()),
+                }
+            }
+            None => {
+                // We need to query the info from the task queue
+                let task_status: TaskStatusResponse = deps.querier.query_wasm_smart(
+                    task_queue.to_string(),
+                    &TaskQueryMsg::TaskStatus { id: task_id },
+                )?;
+                // Abort early if not still open
+                match task_status.status {
+                    TaskStatus::Completed => Err(VerifierError::TaskAlreadyCompleted),
+                    TaskStatus::Expired => Err(VerifierError::TaskExpired),
+                    TaskStatus::Open => {
+                        // If we create this, we need to calculate total vote power needed
+                        let total_power: TotalPowerResponse = deps.querier.query_wasm_smart(
+                            operator_addr.to_string(),
+                            &OperatorQueryMsg::TotalPowerAtHeight {
+                                height: Some(task_status.created_height),
+                            },
+                        )?;
+                        // need to round up!
+                        // NOTE: for now we are using percentage, with oracle-verifier
+                        // this would change
+                        let fraction = Decimal::percent(fraction_percent as u64);
+                        let power_required = total_power.power.mul_ceil(fraction);
+                        let meta = TaskMetadata {
+                            power_required,
+                            status: TaskStatus::Open,
+                            created_height: task_status.created_height,
+                            expires_time: task_status.expires_time,
+                        };
+                        tasks_storage.save_tasks(
+                            deps.storage,
+                            (task_queue, task_id),
+                            meta.clone(),
+                        )?;
+                        Ok(meta)
+                    }
+                }
+            }
+        }
     }
 }
